Added Siesta Implementation
First draft of the implementation of the aiida-common-workflows features in the aiida_siesta.
The ProtocolRegistry collects the input values for each protocol and the methods regarding the protocol information. The SiestaRelaxationInputsGenerator implements the common relaxation features discussed in Bologna and relative methods to handle them. The workchain_relax_protocol.py presents a workchain inside which the protocols are called, the relaxation is performed, and the standard outputs are returned.
Very first implementation! Surely can be improved after discussions.