Add abstract class for relax input generator and workchain
The new `common` module adds two base clases that will serve as the
generic interface for common workflows that perform a structure
relaxation. The `RelaxInputsGenerator` class is the main entry point for
the user, and it should be implemnented by each plugin. Specifically, it
defines the available relaxation types and implements the `get_builder`
that will return a completely prepared process builder for the wrapper
workchain that is based on the `CommonRelaxWorkChain` class. This is the
base class for a wrapper workchain that will expose the relax workchain
of a particular plugin and convert the outputs to those matching the
common output specification.

Each plugin should implement these two classes in a sub folder in the
module `aiida_common_workflows.plugins.relax`. The usage would then be
as follows, for example the Quantum ESPRESSO implementation:

    from aiida.engine import submit
    from aiida_common_workflow.plugins.relax import  QuantumEspressoInputsGenerator
    generator = QuantumEspressoInputsGenerator()
    builder = generator.get_builder(structure, protocol....)
    submit(builder)

This will submit the wrapper workchain with the inputs that were
automatically determined by the input generator. The resulting workchain
will produce outputs with a format and naming convention that is shared
between all implementations.