CI: simplify and improve the `pre-commit` configuration
The `prospector` hook is dropped. The reason is that it is mostly used
to run `pylint`, but we can run that just directly. The `prospector`
tool has given quite a lot of problems regarding dependency
incompatibilies in the past that would cause the CI to fail for
unrelated reasons. In addition, it runs the `pep8` tool whose function
is already fulfilled by `yapf` but worse. For example, its line length
checker does not account for pylint disable statements, giving false
positives.

Furthermore, the order of `yapf` and `pylint` is reversed, with `yapf`
being called first. This prevents `pylint` complaining about incorrect
formatting when `yapf` hasn't had the chance yet to address it.

Finally, `yapf` is directly installed from the remote repository into
the special `pre-commit` virtual environment. This makes it easier to
run multiple projects in the same virtual environment that may be
running different versions of `yapf`. This also allows us to remove it
as a direct dependency in the `setup.json`.