Drop Python 3.5 support officially
Python 3.5 is officially EOL September 23rd 2020, which is in two
months, but most of the big libraries, such as `numpy` have long dropped
support for this minor version. There is little to no reason for this
package to continue supporting it and dropping it will allow us to use
all new features of Python 3.6, such as f-strings and full support for
the `pathlib.Path` module in the `os` interface.