Add entry points for all implementations of `CommonRelaxWorkChain` (#29)
Having entry points for the work chain implementations has various
advantages, such as adding support for `verdi plugin list`. In addition,
a test is added that checks that the work chains can be successfully
imported and that the inputs generator can be retrieved from it through
the `get_inputs_generator` class method.