Implement the common Equation of State workflow
This `WorkChain` employs the available implemenations of the
`CommonRelaxWorkChain` to compute the equation of state of a given
crystal structure.

The work chain spec does not actually expose the inputs of the common
relax workflow that will be used as the executor, because the specific
plugin implementation that is to be used will only become known when the
work chain is launched, i.e., when an instance is created. This prevents
the classmethod define to expose the inputs. Instead, it defines a
dynamic input namespace called `sub_process`. Whichever relax workflow
is to be used, its required inputs are to be passed into this namespace.
This means that the inputs won't be validated at the top level, but once
the EOS workflow submits the sub process. The relax workflow that is to
be used is communicated through the `sub_process_class` input, which
expects its entry point name.

The `EquationOfStateWorkChain` will launch the sub process for the input
structure that is first scaled at a list of scaling factors. These can
either be specified explicitly as `scaling_factors`, or by default they
will be generated by the work chain itself. The number of scaling
factors and the relative increment can be controlled with the inputs
`scale_count` and `scale_increment`, respectively. The EOS workflow will
simply return the relaxed structures and total energies computed by the
common relax workflows.