Add command line interface utility (#34)
This CLI for now has a single command, which can be used to launch the
common relax workflow for one of the plugin implementations. For example:

    aiida-common-workflows launch relax siesta

will launch the implementation of Siesta. By default it will use a
simple silicon crystal structure (that will be retrieved from the
database or created if none can be found) but a specific structure can
be passed using the `--structure` option.

The command will use the inputs generator, retrieved from the workflow
entry point, to determine which calculation engines are required by the
workflow and will attempt to load a code with the correct input plugin.
If no code can be found, the command prints an error.

The `--show-engines` flag can be used to have the command display an
overview of all engines that are required along with a description and
the required code plugin.

Note that I removed `aiida_common_workflows.workflows.relax.workchain`
from the `__all__` in `aiida_common_workflows.workflows.relax` because
the `aiida.engine` import in the former, forces a profile and database
backend to be loaded, which is slow. Since some of the constants defined
in `aiida_common_workflows.workflows.relax.generator` are required to
define the command options, such as the available relax types, this
requires loading this module, but we don't want it to require loading
the profile, since it will make tab-completion slow and unusable.