CLI: allow to define resources for each engine step in relax launch
Previously, the resources, i.e. the number of machines and wallclock
seconds, to be requested was hardcoded for all engine steps. Now these
are configurable through the `--wallclock-seconds` and
`--number-machines` options. There is validation to make sure they are
of equal length and match the number of engine steps for the common
relax workflow that is selected.

In addition, the code is now passed by its full label, which will make
sure it can be loaded even in case of ambiguity because there are
multiple codes with the same label configured for multiple computers.