Move check on "previour_workchain" in the abstract class
The check on the value passed to the "previour_workchain" argument
of `get_builder` is general enough to be moved in the `RelaxInputsGenerator`
abstract class.
This implies the call to the `super` inside `get_builder` of the
plugin specific implementations of the input generator.
This has been added only for siesta at the moment.