Implement the common Equation of State workflow (#35)
This `WorkChain` employs the available implemenations of the
`CommonRelaxWorkChain` to compute the equation of state of a given
crystal structure. The relax workflow that is to be used is communicated
through the `sub_process_class` input, which expects its entry point name.

The `EquationOfStateWorkChain` will launch the sub process for the input
structure that is first scaled at a list of scaling factors. These can
either be specified explicitly as `scaling_factors`, or by default they
will be generated by the work chain itself. The number of scaling
factors and the relative increment can be controlled with the inputs
`scale_count` and `scale_increment`, respectively. The EOS workflow will
simply return the relaxed structures and total energies computed by the
common relax workflows.

The work chain spec does not actually expose the inputs of the common
relax workflow that will be used as the executor, because the specific
plugin implementation that is to be used will only become known when the
work chain is launched, i.e., when an instance is created. This prevents
the classmethod define to expose the inputs. On top of that, the
workchain itself will have to call the inputs generator of the plugin
implementation that is to be used in order to generate the inputs. One
would think this could be done once before calling the EOS workchain and
but that is not possible. Inside the workchain, the inner workchain has
to be called multiple times with a different structure, but the outer
workchain cannot now where in the input hierarchy the input structure
should be passed. Not all inner implementations necessarily expose the
structure port at the top level. Therefore the outer workchain has to
call the generator.

Another reason is that the input generator is potentially dependent on
a previously run workchain, through the `previous_workchain` keyword.
This is because for an EOS to make sense, certain implementations will
need to keep certain variables constant across the various calculations.
This again means the workchain will need to call the generator inside
the workchain and we cannot simply generate the inputs outside the
workchain and pass this is.

This all means that the inputs that are supposed to be passed to the
generator, also need to be passed to the workchain itself. For now we
add a port `generator_inputs` that takes a dictionary of the values that
are to be passed to the input generators. Maybe we want to make this a
`non_db` portnamespace in the future, since it doesn't really need to be
part of the provenance, just as normally calling the input generator is
not part of the provenance, which is fine.

The workchain still defines a dynamic input namespace `sub_process`.
This is completely optional but any input passed here will override the
inputs created by the generator. This allows a caller to override
certain inputs should they want to. It is not exactly clear yet how this
will work and if it will be practical.