`EosWorkChain`: make `generator_inputs` non-db namespace (#50)
Turning the `generator_inputs` input from a normal port expecting a
`Dict` node into a namespace where all ports are non-db, brings a few
advantages:

 * Validation of values at launch time instead of within the workchain
 * No longer need to serialize values, such as `RelaxType.ATOMS`
 * Benefit from spec introspection, help strings and validation

It also allows the workchain code itself to be easier as now it does not
have to deserialize the content and can simply pass the entire namespace
as keyword arguments to the inputs generator.

For now, the `calc_engines` input is not further decomposed in
namespaces but is defined as a single dictionary. We could think of
expanding this further and adding more validation.