Add previous workchain to fleur relax
the builder now extracts flapw parameters from the last fleurinpdata
of the last scf workchain it finds. So far we do not recycle kpoints.
Also fix and slimm the protocols by merging with defaults.
So far there is just a precision difference for the charge distance
not yet for basis cut offs or kpoint densities.