Fleur fix (#52)
* Update fleur relax to common interface

Before inpgen was a second entry in the relax calc engine entry
which was not conform to convention and due to a miss understanding.
this is now fixed with inpgen becoming an independent entry

* Protocols
Added protocol.yml file containing all the protocols
for fleur. Slim the protocols by merging with defaults.
There is just a precision difference for the charge distance 
and kpoint densities not yet for basis cut offs.

* Previous workchain
The builder can now work with a code node given in any form, 
making the cli launch relax fleur command work.
The builder now extracts flapw parameters from the last fleurinpdata
of the last scf workchain (previous workchain) it finds. 
So far we do not recycle kpoints.
Works now with aiida-fleur v.1.1.2.