Introduce the InitialMagnetizationManager.
The `initial_magnetization` argument of `get_builder` accepts
a list containing the spin polarization of each site.
A manager has been created to facilitate the management
of this `initial_magnetization`.
The idea is to gather here useful methods that developers
can call inside their `get_builder` to process the `initial_magnetization`
format agreed on the common API.
For the moment it has only a method to modify the structure kinds
in order to accomodate initializations where sites of the same king
have different magnetizations.