`RelaxInputsGenerator`: add arguments to specify magnetic/electronic type (#64)
New arguments are added to the `get_builder` method:

 * `electronic_type`
 * `spin_type`
 * `magnetization_per_site`

The first specifies the electronic type and is defined through the
`ElectronicType` enum which has values `AUTOMATIC`, `INSULATOR` and
`METAL, which is the default. The `AUTOMATIC` option does not have to be
supported by all plugins.

The `spin_type` defined by the `SpinType` enum can take the value `NONE`
for unpolarised calculations and `COLLINEAR`, `NONCOLLINEAR` and
`SPIN_ORBIT`. For now only `NONE` and `COLLINEAR` needs to be supported.

Finally, in the case of a magnetic calculation, the initial magnetization
can be defined by `magnetization_per_site`. If it is not defined, the
plugin needs to automatically define a suitable magnetization.