`RelaxInputsGenerator.get_builder`: use keyword arguments only syntax
All arguments for the `get_builder` method are marked as keyword
argument only, except for `structure`, `calc_engines` and `protocol`.
This will make the interface more stable with respect to changes we
might have to apply in the future. If we want to add new arguments,
change their order etc. by making them keyword only, calls of this
method will not be affected.

The choice to make `structure`, `calc_engines` and `protocol` positional
arguments is that they are always required and don't have a default.

In addition, the argument `relaxation_type` is renamed to `relax_type`
which makes more sense given that the enum that it takes is called
`RelaxType`.