`QuantumEspressoRelaxInputsGenerator`: implement missing features
Implements the following new arguments for `get_builder`:

 * `spin_type`
 * `relax_type`
 * `electronic_type`
 * `magnetization_per_site`
 * `previous_workchain`

Additionally add following changes:

 * Make the `relaxed_structure` output optional
 * Add the optional `total_magnetization` output
 * Change `initial_magnetization` to `initial_magnetic_moments`

The protocol builder interface in `aiida-quantumespresso` changed the
`initial_magnetization` input to `initial_magnetic_moments` because the
latter is more universal and easier to reason about. On top of that,
this way one does not have to have knowledge about the pseudo potentials
because the number of electrons in it determine the magnetization for a
given magnetic moment. This was causing a catch 22 in the input protocol
builder method.