`RelaxInputsGenerator.get_builder`: make `protocol` keyword and optional
The protocol argument was required, even though the class underlying the
generator, `ProtocolRegistry`, makes specifying a default protocol
required for any subclass. Requiring a caller to explicitly define the
protocol when a protocol exists, doesn't make any sense.

That is why the `protocol` argument's default is set to `None` and in
the implementations, the `get_default_protocol_name` method should be
called in case no specific protocol is defined.

The argument is also made keyword only, to keep in line with all other
arguments that are not required. This only leaves `structure` and
`calc_engines` as the arguments that can be specified positionally.