CLI: allow to pass structure file to `-S` structure option (#81)
The value passed to the flag will first be interpreted as a node
identifier and if that fails, it will be interpreted as a file path. If
`ase` is installed it will attempt to parse a `StructureData` from it
and return it. The option performs a check on the database to see if the
exact same structure already exists in which case that is returned
instead of creating a new node.