Use `click.types.Choice` and add unit tests
By using the `click.Choice` type as base class, the help string
automatically lists the values of default structures that are available.
I have changed the name to use the element or formula, which is shorter
to type, since they are not autocompleted. I have also added a
parametrized test to make sure that the structures are loaded from the
correct cif if one of the default structure labels is specified.