CLI: add choice of default structures to `--structure` option (#84)
The `-S/--structure` option already supports loading a structure from
a `StructureData` identifier and directly from file that can be parsed
as a structure by `ase`. Here we add also a few keywords that map to
structure files that are included with the repository: Al, Fe, GeTe, Si,
NH3-pyramidal, NH3-planar and H2. If one of these strings is passed the
structure is loaded from the corresponding file in the `common/data`
directory. These structures are used for the proof-of-concept that is
demonstrated in the first paper.

Finally, the requirement `ase==3.19` is pinned because `3.20` has a bug
and fails to parse `Al.cif` and `Fe.cif`.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>