EoS workchain: add validators.
Added validators for new optional keys allowed in the `generator_inputs`
dictionary (input of the EoS workchain). They reflect the last changes
in the `get_builder` signature, menaning the addition of `spin_type`,
`electronic_type` and `magnetization_per_site` arguments.

Also moved the check on the allowed relaxation types to a dedicated validator.