CLI: make launch commands allow to specify codes manually (#80)
So far, the workchain launch commands attempted to automatically load
the required code based on the information provided by the selected
plugin. However, in the case where multiple codes exist in the database
for a given calcjob plugin, a random one would be chosen, which might
not always be the desired one.

The option `-X` is added that allows a user to explicitly define the
codes that are to be used. If a code is not explicitly defined, the
command will still retrieve an arbitrary one from the database that has
the correct calcjob plugin defined.