CLI: allow to specify float precision for plot commands (#108)
The `plot-eos` and `plot-dissociation-curve` allow to print the results
in a table to stdout instead of plotting them, but the floats are
printed with default precision which is not always enough to be useful.
The `-p/--precisions` flag allows a user to specify the exact precision
to use for the floats in each column that is printed when using the
`--print-table` option. If less values than columns are specified,
defaults are used for the remainder. Any surplus values that are
specified are simply ignored.