CLI: add `--previous-workchain` to `launch relax` command
The option allows to pass a node representing an already completed
common relax work chain of a particular plugin which should be used by
the builder to determine the inputs. The protocol's `get_builder` will
validate that the node actually corresponds to one run by the plugin
that is requested to run.

Note that the same option is not added for the other launch command
`eos` and `dissociation-curve` because there the workchain will itself
run a single workchain first that is subsequently passed to the
generator of all other workchains.