Updated VASP interface
In order to comply with recent upgrade of specifications, we here add the required translation to the VASP workchain stack for performing relaxations, simple equation of state calculations and a dissociation example. Since VASP requires a dedicated version for non-collinear calculations we do not include this or spin-orbit as an option at this point. We also do not make use of our convergence workchain in order to keep it as close to code as possible.