Fix assigning magnetic moments
Currently the `QuantumEspressoRelaxInputGenerator` fails to assign the
correct magnetic moments based on the `magnetization_per_site` input.
The reason is that the `zip` generator is exhausted by the conversion
into a `set` when comparing the kinds of the structure with the mapping
of the kinds to the magnetic moments provided. Hence, the `dict`
assigned to the `initial_magnetic_moments` variable will be empty.