Add `create_magnetic_allotrope` function
Currently running antiferromagnetic configurations with Quantum ESPRESSO
requires that the user first creates a new structure with the necessary
kinds. This is not trivial if the user has no experience in using AIIDA.

Here we add the `create_magentic_allotrope` function, largely based on
the CP2K variant `tags_and_magnetization`, but with some minor
adjustments to make it work properly for Quantum ESPRESSO. This function
is called when the `get_builder` method of the
`QuantumEspressoRelaxInputGenerator` notices that the provided
`magnetization_per_site` would require new kinds in the provided
`structure`.

Note that a disadvantage of this approach is that a new (possibly
duplicate) `StructureData` is created every time the `get_builder` is
called. However, this would only be stored if the returned builder is
actually run/submitted. Moreover, for more complex structures/magnetic
configurations, this would lead to kind names with 4 characters, which
is currently not supported by Quantum ESPRESSO (max 3).