Relax workflow: give all implementations same suffix
The common suffix used so far for all implementations of the common
relax workchain was `RelaxWorkChain`. However, given that many plugins
already use this suffix for their own specific relax workflows, it was
difficult to distinguish these workchains from the one with the common
interface from `aiida-common-workflows`. Therefore, we add the term
`common` explicitly into it and make the suffix `CommonRelaxWorkChain`.