QE: bug fixes for `QuantumEspressoCommonRelaxInputGenerator` (#168)
There was a bug for the `threshold_forces` and `threshold_stress`
keyword arguments that went undetected because there were no unit tests
for the generator. The reason is that running the generator presupposes
certain data to be present in the database, in this case, the SSSP
pseudo families. These are not trivial to create.

Here, we add fixtures to setup these requirements which then allow to
define simple unit tests for the various keyword arguments that the
input generator's `get_builder` method supports. Note that there is no
test for `reference_workchain` since this requires the mocking of a
completed `QuantumEspressoCommonRelaxWorkChain` which is not trivial.