`CommonRelaxInputGenerator`: allow str for enum keyword arguments (#182)
Currently, the `get_builder` method required an Enum instance to be
passed for the `electronic_type`, `relax_type` and `spin_type` keyword
arguments. This requires the user to import the corresponding enums
which forces not just to add a line of code but also for them to know
where to import it from. From a user's perspective it is easier to just
be able to pass the string value as well.

The signature is updated to also accept a string. The downside is that
the implementation now first has to check the type. To prevent each
implementation having to do this, we add as much as possible of the
validation in the base class, however, it is not possible to change the
keyword argument values themselves. So we cannot ensure in the base
class to convert any str types to the corresponding enum. This forces
the implementations to still perform this check and conversion
themselves, but at least they know that whatever value is received it is
a valid and supported enum value.