CLI: Move `plot` commands to separate top-level command (#187)
Move the `plot` commands, currently subcommands of the top-level launch
command, to a separate top-level command called `plot`. This is more
intuitive, since the `launch` commands (as the help describes) are for
"launching" the work chains.

Also sort plot results by volume/distance for the eos/dissociation. This is not
only important for testing, but makes more sense from a user perspective.