CLI: Add -o/--output-file option to plot commands
Add an extra option to the plot commands that allows the user to save
the plot or table to a file instead of showing/printing the results.

Also refactor the plotting code for the dissociation curve to a new
module (common/visualization.dissociation.py), and return the plot for
the plotting functions in the visualization package so they can be
either shown or saved to file in the CLI commands.