Composite: validate `generator_inputs` for EOS and DC workchains (#189)
Currently, the inputs in the `generator_inputs` are not validated on
instance creation of the composite workflows. This means that even if
they are invalid, the workchain will actually be created and stored in
the database and will not fail until `get_builder` of the relevant
inputs generator is called.

To prevent this, the `generator_inputs` are validated through the spec
by calling the `get_builder` method in the `validate_inputs` validator
that validates the entire input namespace of the composite workchain.

In addition, the `valid_type` of the composite workchains' input ports
for the electronic, spin and relax type is updated to also allow a
string type value, in addition to the enum type. The input generators
were recently updated to support the string type as well, but the
composite workchain specs were forgotten.

Finally, the `generator_inputs` namespaces were marked as `dynamic`
accepting any input even if no explicit port is defined for it, but this
is really not necessary and in this case can hide incorrect inputs that
are passed by mistake, so the attribute is removed.