`AbinitCommonRelaxInputGenerator`: use new unit feature of `aiida-pseudo`
Since `aiida-pseudo==0.6.0` families can store recommended cutoffs in
different units and when requesting them with `get_recommended_cutoffs`
the desired unit can be specified and the method will perform the
conversion. The `AbinitCommonRelaxInputGenerator` was assuming that the
units would always be returned in eV, which used to be the case for
older versions of `aiida-pseudo` but now the cutoffs will be specified
for the unit with which they were defined, unless explicitly asked for
another unit.