CI: install the package normally without editable flag
The `-e` flag for `pip install` installs the package in editable mode,
meaning that the repository is simply symlinked into the virtual
environment such that changes to the code are picked up. This is useful
during development, however, for testing it is not necessary, but worse,
it can hide bugs with the definition of the package. For example,
non-python files that have to be shipped with the package, for example
yaml configuration files, need to be properly defined in `MANIFEST.in`
or else they will be missing in the distribution. By using an editable
install, these mistakes won't be noticed, since these files will always
be available directly from the repository.