CI: add testing for common relax input generator and input spec
Here we add basic unit tests for the `CommonRelaxWorkChain` for all
plugin implementations. It consists of two main parts essentially:

 * Call `InputsGenerator.get_builder`
 * Submit the builder returned by `InputsGenerator.get_builder`

The first test is performed multiple times: first with minimal required
inputs, i.e., just `structure` and `engines`. Then it is called with a
specific value for `electronic_type`, `relax_type` or `spin_type`. The
values that are passed are taken by introspecting the input generator
class itself that declares which values it supports. In this way we
check that those declarations are correct.

For the call to `get_builder` with default arguments, the returned
builder is actually submitted. This will actually force the
implementation of the `CommonRelaxWorkChain` class to be instantiated
which in turn will validate the generated builder against its input
specification. If this test passes, it means that the workchain can at
the very least be submitted and successfully created. This is no
guarantee that it would then of course successfully finish, but this
integration test is really complicated to setup in the general case, and
so is outside of the scope of these unit tests.

The difficulty in this commit is that certain input generators rely on
resources being preinstalled, for example pseudopotential families. This
is the case for Abinit, Siesta and Quantum ESPRESSO. The relevant
fixtures are added to configure those requirements that the input
generators will expect to be present.

Note that the `test_submit` tests for is skipped for all plugins since,
for an unknown reason, these fail now and again on Python 3.6 and Python
3.7. The failure is related to the process failing to connect to
RabbitMQ but it seems completely unrelated to the specific test but
merely a question of the quantity of tests needing to establish a
connection.