`DissociationCurveWorkChain`: add `electronic_type` input (#200)
This input was missing from the spec but it was being used by the CLI
command to launch it causing an exception. This went unnoticed because
it wasn't tested. Tests are added for the launch commands:

 launch relax
 launch eos
 launch dissociation-curve

using the default options. This will give some coverage ensuring that
the launch commands work properly.