Update NWChem generator for molecular calculations
Changes to make the protocol more robust in the case of a molecular calculation.
NWChem's plane wave code currently allows the simulation cell to be specified in
two different ways. While both will be valid in the sense that a calculation
will be run, using the traditional way of adding a cell will force the geometry
optimisation module, "Driver" to use cartesian coordinates for the optimisation,
rather than internal degrees of freedom. While this seems harmless, testing has
shown that highly converged geometries are not possible as the optimisation will
get stuck by making vanishingly small steps, while the forces are small but larger
than the target tolerances.