Adding a --engine-options option to the CLI
This allows to specify a dictionary (as a JSON string) to
add more content to each of the engines.

Example usage and usecase:
`--engine-options='{"account": "mr0"}'`
when, in order to run, you are required to specify an account.

This to me seems a still general enough approach, and important
(without it I cannot run so the CLI is not very useful to me).