Fixtures: fix bug in pseudopotential node generating fixtures
The fixtures are constructing a pseudopotential node of various types
by writing the content of a pseudo in the respective format to a
temporary file but then passing the stream to the constructor.
However, the stream was just written to and so the pointer will be at
the stream which will cause the constructor of the node to read an
empty string of bytes.

Since we are using a stream in memory, we don't actually need a file on
disk at all, so the temporary file is replaced by a simple `io.BytesIO`
stream which solves the problem. Note that this does oblige us to
explicitly set the filename since otherwise the default `file.txt` would
be used, but some of the tests expect a different name.