Tests: fix generator tests for Siesta plugins
As of `aiida-siesta==1.2` the `options` for the calcjob in the base
workchain have to be passed as an individual `Dict` node on the top
level of the workchain input space because the inputs of the calcjob are
not exposed into a namespace. The options port includes a validator that
makes the `max_wallclock_seconds` option required. This broke the tests
here since that option wasn't specified in the inputs.