CLI: add `--engine-options` parameter to specify calcjob options (#205)
This allows to specify a dictionary (as a JSON string) to
define specific options for each of the engines. Example usage:

    --engine-options='{"relax": {"account": "mr0"}}'

when, in order to run, you are required to specify an account.
Note that the engine to which the options should be applied needs
to be specified. By requiring this key, different values for the same
option can be specified if a plugin has multiple engines for the workflow.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>