Implement the concept of input generator specifications
The refactoring of the input generators for common workflows in this
commit comes from the following two needs:

 1. The exact signatures of the `get_builder` method of input generators
    needs to be customizable, as it has become clear that certain
    implementations need additional keyword arguments that may not be
    general enough to fit all implementations.

 2. The signature of the `get_builder` method of input generators needs
    to be introspectable. That is to say, it needs to be possible to
    programmatically determine which arguments it accepts, what the
    valid types for those arguments are, and additional information,
    such as help strings.

These requirements very closely match that of the `Process` base class
of `aiida-core` itself. It needs to allow subclasses to customize their
specification with respect to inputs and outputs. This functionality is
implemented through the concept of a process specification that build up
their input and output namespaces through `PortNamespace`'s and `Port`'s.

Instead of reinventing the wheel, we repurpose this concept and create a
base `InputGenerator` class. This base class defines the `define`
method, just like a `Process`, which gets an instance of the new class
`InputGeneratorSpec`. This is the analogue of the `ProcessSpec` and
allows to define the input specification of the input generator by
creating a, optionally nested, namespace of `InputGeneratorPort`s. These
define the keyword arguments that the `get_builder` method of the input
generator will accept.

By using the process specification functionality of `aiida-core`, it
becomes easy to define specification for input generators whose
validation is fully automated and whose specification is
programmatically introspectable. This will allow clients to
automatically build interfaces that interact with the input generators,
as well as automatically generate human-readable documentation.