Docs: add a Sphinx extension to auto document input generators
With the recent change of having `InputGenerators` build the signature
of their `get_builder` method dynamically through an `InputGeneratorSpec`,
just as a `Process` specifies its input signature through a `ProcessSpec`,
we can now introspect the signature fully programmatically. This in turn
allows us to generate any visual representation of said signature.

Here, we add a Sphinx extension that generates HTML from the specification
of a input generator for a common base workflow. This makes it very easy
to automatically generate rich, complete and consistent documentation for
all of the implementations.

The Spinhx extension is immediately put to use by adding the docs for all
the implementations of the common relax workflow.