Fix multiple inheritance
No need for constructor in `CommonRelaxInputGenerator` if it doesn't add
any functionality. Base class constructors will be called automatically.

Also make `process_class` an explicit keyword argument for `InputGenerator`.
Finally, reverse the order of subclasses `CommonRelaxInputGenerator`
such that the `process_class` argument is only passed to the generator
class and not the `ProtocolRegistry`.