Avoid deepcopy of nodes in general
The problem of deepcopying the inputs of an input generator does not affect the Data
nodes only. Also other nodes like `ProcessNode` should not deepcopied
(actually in that case the code will crash trying to deepcopy).
So the deepcopy is prevented for all `Node`s.
Also eliminated the distintion betweeen stored and unstored nodes.
An implementation can not know in advance if the passed Data node is
stored or not. Therefore every implementation should not modify
nodes at all! No need to deepcopy nodes that are not stored!