`EquationOfStateWorkChain`: fix `reference_workchain` use
In the current implementation, the `reference_workchain` is always
passed to the `get_builder()` method of the input generator in the
`get_sub_workchain_builder()` method. However, in case the
`reference_workchain` is `None`, this leads to an error with
the new input generator specification based on the spec/ports.

Here we change the `get_sub_workchain_builder()` method to only pass the
`reference_workchain` to the `get_builder()` method in case it is not
`None`.