`get_builder`: do not deep-copy nodes (#243)
The `kwargs` passed to `get_builder` of an input generator are deepcopied to ensure
that any changes to them are not applied to the original inputs, which may have to be
reused. However, nodes should not be deepcopied. For starters, for `ProcessNode`s
this will except as it is not allowed, but for `Data` nodes a clone will be created, which
is not what we want however as it loses provenance. The fix is to have a custom
recursive function that deepcopies the `kwargs` but skips all nodes.