Introduce the `RelaxAndBandsWorkChain`
It is a code agnostic workchain that runs first a relaxation and
subsequently a bands calculation. It allows the use of seekpath
to obtain the path in k space where to calculate the bands
(actually this is the default behaviour). In such case an extra
scf is run after the relaxation because seekpath might change
the structure. The workchain also allows to use one
code for relaxation and another for bands. An extra
scf step is needed on the relaxed structure with the plugin
selected for relaxation. The handling of this case is fragile
at the moment and will be improved.
The implementation requires to expose the inputs of the input
generator for relax and of the input generator for bands. A system has
been set up to deal with the fact that a workchain requires Data
Nodes as inputs, while the inputs generators accept python types.