`InputGenerator`: ensure the `_spec` attribute is subclass specific (#248)
The `spec` classmethod was using the global `getattr` method to check
whether the class already had the `_spec` class attribute defined, to
determine whether a new generator spec instance needed to be
instantiated and assigned to that attribute. However, this method will
go up the MRO of the class to see if any of its base classes already
have the attribute. This is a problem because the `_spec` attribute
should be subclass specific. If not, calling `spec()` on a base class
generator will lock the spec for all its subclasses as well, and so when
they call the spec, their own define method is not applied and they will
have the spec definition of the base class.

The solution is to replace `getattr` with `cls.__getattr__` which does
not subclasses and so remains subclass specific.