Add the `CommonBandsWorkChain` (#252)
This base common workchain interface can be implemented by plugins that
can compute an electronic band structure from a previously computed
charge density.

Since the charge density will have been computed through the common
relax workflow, the input generator for this bands workchain does not
specify the typical protocol or any other protocol modifying arguments.
The only accepted and required inputs are:

 * a list of kpoints where to calculate the bands
 * an engine dictionary containing the code and computational resources
 * a "parent_folder" from which all the other inputs should be obtained.

This implementation requires a small change on the abstract class for
the input generator as that used to subclass `ProtcolRegistry` which
effectively requires all implementations to define a protocol. With the
common bands workchain we see that this requirement shouldn't be there,
so `ProtocolRegistry` is removed from the base generator class and mixed
in directly into the `CommonRelaxInputGenerator` instead.