Expose the possibility to select inputs of the second relaxation.
Once an automatic generation of the kpoints path for bands is
performed by SeeKpath, this might refere to a conventional
structure that is different from the input one. For this reason it is necessary
to run a further scf before calculating bands.
The inputs for the further scf are now exposed. They are the inputs of a
`CommonRelaxInpuGenerator`, but where the default relaxation_type is NONE and
all the other defaults have been removed.
If not defined by user, the same inputs of the first relaxation are used
(except the new default for relaxation_type)